requirements.txt
asgiref==3.8.1
cachetools==5.5.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-auth==2.40.1
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
